backend/requirements.txt
Flask
Flask-RESTX
Flask-JWT-Extended
Flask-Cors
Flask-SQLAlchemy
PyMySQL
Werkzeug
Pillow